Update button.mdx
Adding context about when to use a primary button.  In DS, "Page" is a bit generic.  If a page is a form, then a CTA is completely appropriate.  But, if a page is just a page (i.e. a non-form visualization page, as are many in a product), a CTA is often not warranted because the 80% use case for most pages is visualization and we don't need to or want to push a user down an action path, distracting them from the visualization.  In cases like these, a secondary button should be used (default buttons often do not work because they are text and have no affordance in this kind of context).

Also adding documentation about disabled buttons.  In general, disabling controls is a bad practice and should be avoided.  Buttons are the control that is most commonly disabled, so I am adding this documentation here.  But, because most grommet components support disabling, it would be good to have a general statement somewhere that indicates 1) disabling controls should be avoided -- and why; 2) don't show controls that simply do not apply; 3) use progressive disclosure instead of disabling controls; 4) active submit buttons should perform form validation and emit an error message if the form cannot be submitted; 5) if it is absolutely necessary to show a disabled control, a message explaining why it is disabled, along with info about how to enable it must be attached to the disabled control.