Update icons.mdx
We have too many icons in grommet and many of them are very obscure.  We really ought to be defining a core set of icons that can be used without needing any textual assistance. Barring that, we should give better guidance (as in this request) about only using commonly understood icons and discouraging use of obscure icons.