Update lists.mdx
This is a quick fix for the inconsistency between the lists guidance and a list example found on menu page.

After discussion with designers, we decided to not restrict the amount of lines in the list without good reasoning. We agreed that as long as the idea of "list should be short and not have too much information" is conveyed, the line restrictions are not needed, hence why I am deleting this section.

There's also the need of reviewing this page in full, since it's outdated, but that needs and issue to track the work.